Remove permission checking from domain
Building a confirm delete page revealed a flaw in the design of having
domain check permissions. A user could have permission to view another
user's book but not to delete it. But the confirmation page would only
be using domain.GetBook.

More abstractly, it needs to be possible to check the permission of an
action without attempting to perform it.

The requireSameSessionUserAndPathUserHandler middleware is already
protecting books so there is no loss of security by this change.